+++ frontend/src/WebUI/Forms/ClassifyFilesForm/ClassifyFilesForm.stories.js
@@ -9,7 +9,9 @@ import { createStyles } from '@material-ui/core';
 const categoryName = 'AssembledComponents/Forms';
 
 storiesOf(categoryName, module).add('ClassifyFilesForm', () => {
-  const [state, setState] = useState({});
+  const [state, setState] = useState({
+    files: [],
+  });
   const onChange = ({
     name,
     value
@@ -19,12 +21,14 @@ storiesOf(categoryName, module).add('ClassifyFilesForm', () => {
       [name]: value,
     })
   };
+
   const classes = createStyles({
     textField: {
       width: 'auto',
       height: 20,
     },
   })
+
   const i18n = {
     classifyFiles: {
       title: 'Formulario para clasificar una evaluación',
@@ -35,6 +39,22 @@ storiesOf(categoryName, module).add('ClassifyFilesForm', () => {
       sheetId: 'Google Sheet ID',
     },
   };
+  
+  const options = [
+    {
+      label: 'SCANS/IIC2333/p1-mt-2019-2',
+      value: 'SCANS/IIC2333/p1-mt-2019-2',
+    },
+    {
+      label: 'SCANS/IIC2333/p2-mt-2019-2',
+      value: 'SCANS/IIC2333/p2-mt-2019-2',
+    },
+    {
+      label: 'SCANS/IIC2333/p3-mt-2019-2',
+      value: 'SCANS/IIC2333/p3-mt-2019-2',
+    },
+  ];
+
   const defaultProps = {
     config: config({
       onChange,
@@ -42,6 +62,8 @@ storiesOf(categoryName, module).add('ClassifyFilesForm', () => {
       i18n,
       classes,
       defaultValue: {},
+      setState,
+      options,
     }),
     title: 'Formulario para clasificar una evaluación',
     i18n: {
